Initial refactoring; Extract helpers method(start, bug, log_print) in helpers.py
